Fix split package in analysis-smartcn plugin (#78045)
The analysis-smartcn plugin reuses server the server package name for
analysis. This commit moves the plugin implementation to use a single
package name, o.e.p.analysis.smartcn
