+++ plugins/analysis-smartcn/build.gradle
@@ -25,11 +25,3 @@ restResources {
 tasks.named("dependencyLicenses").configure {
   mapping from: /lucene-.*/, to: 'lucene'
 }
-
-tasks.named('splitPackagesAudit').configure {
-    // Lucene packages should be owned by Lucene!
-  ignoreClasses 'org.elasticsearch.index.analysis.SmartChineseAnalyzerProvider',
-    'org.elasticsearch.index.analysis.SmartChineseNoOpTokenFilterFactory',
-    'org.elasticsearch.index.analysis.SmartChineseStopTokenFilterFactory',
-    'org.elasticsearch.index.analysis.SmartChineseTokenizerTokenizerFactory'
-}
