+++ plugins/analysis-smartcn/src/test/java/org/elasticsearch/plugin/analysis/smartcn/SimpleSmartChineseAnalysisTests.java
@@ -6,11 +6,13 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.plugin.analysis.smartcn;
 
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.Index;
+import org.elasticsearch.index.analysis.TokenizerFactory;
 import org.elasticsearch.plugin.analysis.smartcn.AnalysisSmartChinesePlugin;
+import org.elasticsearch.plugin.analysis.smartcn.SmartChineseTokenizerTokenizerFactory;
 import org.elasticsearch.test.ESTestCase;
 import org.hamcrest.MatcherAssert;
